feat: refactor toast for hooks
