--- src/components/Toast/index.ts
@@ -1,4 +1,0 @@
-import JoinRequestByToast from "@/components/Toast/JoinReqeustBy";
-import UserDisconnectedToast from "@/components/Toast/UserDisconnected";
-
-export { JoinRequestByToast, UserDisconnectedToast }
