--- src/env.ts
@@ -1,2 +1,0 @@
-const VITE_BASE_URL = "https://code-sync.net/api/";
-export default VITE_BASE_URL;
