+++ src/pages/conversation/ConversationPage.tsx
@@ -1,13 +1,16 @@
 import { CodeEditor, CodeSplitEditor } from "@/components/CodeEditor";
 import { LeftGNB, TopGNB } from "@/components/GNB";
-import { JoinRequestByToast, UserDisconnectedToast } from "@/components/Toast";
-import { SocketJoinRequestBy } from "@/components/Toast/JoinReqeustBy";
-import { SocketUserDisconnected } from "@/components/Toast/UserDisconnected";
 import {
   ResizableHandle,
   ResizablePanel,
   ResizablePanelGroup,
 } from "@/components/ui/resizable";
+import useJoinRequestByToast, {
+  SocketJoinRequestBy,
+} from "@/hooks/Toast/useJoinReqeustBy";
+import useUserDisconnectedToast, {
+  SocketUserDisconnected,
+} from "@/hooks/Toast/useUserDisconnected";
 import { PrChangedFileInfo, prInfoStore } from "@/stores/github.store";
 import { socketStore } from "@/stores/socket.store";
 import { useEffect } from "react";
@@ -16,23 +19,25 @@ const ConversationPage = () => {
   const { prInfo, selectedFile, prChangedFileList, setSelectedFile } =
     prInfoStore();
   const socket = socketStore((state) => state.socket);
+  const { onToast: onJoinRequestByToast } = useJoinRequestByToast();
+  const { onToast: onUserDisconnectedToast } = useUserDisconnectedToast();
 
   useEffect(() => {
     if (!socket) return;
 
     socket.on("join-request-by", ({ data, message }: SocketJoinRequestBy) => {
-      JoinRequestByToast({ data, message });
+      onJoinRequestByToast({ data, message });
     });
     socket.on("user-disconnected", (data: SocketUserDisconnected) => {
-      UserDisconnectedToast(data);
+      onUserDisconnectedToast(data);
     });
 
     return () => {
-      socket.disconnect();
+      // socket.disconnect();
       socket.off("join-request-by");
       socket.off("user-disconnected");
     };
-  }, [socket]);
+  }, [socket, onJoinRequestByToast, onUserDisconnectedToast]);
 
   const fileOnclick = (selectedFile: PrChangedFileInfo) => {
     setSelectedFile(selectedFile);
